VideoTranscriptor/requirements.txt
openai-whisper
yt-dlp
pydub
opencv-python
boto3
mediapipe
tqdm
